implemented soliciteMove() in HumanVHuman.java, along with adding inevitableDrawEval.java and test for it.
